Refactor the test cases class names. Added test case: VerifyRightClickOptionsOfAnEditedAndRequestPublishPageUnderPageStructureUsingAuthorUser To Cover Crafter CMS Integration Test spreadsheet test case ID:22 Modify the testing.xml and add the created test case. Added test case: VerifyRightClickOptionsOfAnEditedAndRequestPublishPageUnderPageStructureUsingDeveloperUser To Cover Crafter CMS Integration Test spreadsheet test case ID:23 Modify the testing.xml and add the created test case. Added test case: VerifyRightClickOptionsOfAnEditedAndRequestPublishPageUnderPageStructureUsingPublisherUser To Cover Crafter CMS Integration Test spreadsheet test case ID:24 Modify the testing.xml and add the created test case. Added test case: VerifyRightClickOptionsOfAnEditedAndRequestPublishPageUnderPageStructureUsingReviewerUser To Cover Crafter CMS Integration Test spreadsheet test case ID:25 Modify the testing.xml and add the created test case.
