+++ src/test/java/org/craftercms/studio/test/cases/sitedropdowntestcases/VerifyRightClickOptionsOfAPagesUnderPageStructureUsingDeveloperUser.java
@@ -1,0 +1,705 @@
+package org.craftercms.studio.test.cases.sitedropdowntestcases;
+
+import org.testng.annotations.Test;
+import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
+
+import java.util.LinkedList;
+import java.util.List;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.craftercms.studio.test.cases.BaseTest;
+import org.openqa.selenium.By;
+import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.WebElement;
+
+/**
+ * 
+ * 
+ * @author Juan Camacho A
+ *
+ */
+//Test Case Studio- Site Dropdown ID:11
+public class VerifyRightClickOptionsOfAPagesUnderPageStructureUsingDeveloperUser
+		extends BaseTest {
+
+	private String userName;
+	private String password;
+	private String siteDropdownElementXPath;
+	private String pagesTreeLink;
+	private String pagesTree;
+	private String homeContent;
+	
+	private String newUserFirstNameId;
+	private String newUserLastNameId;
+	private String newUserEmailId;
+	private String newUserUserNameId;
+	private String newUserPasswordId;
+	private String newUserPasswordVerificationId;
+	private String newUserUserNameCreatedXpath;
+	private String crafterLogo;
+	private String expandPagesTree;
+	private String createSiteButton;
+	private String adminConsole;
+	private String siteconfigGroupsOption;
+	private String addTouserIframe;
+	private String editDeveloperGroupOption;
+	private String groupsAddNewMembersCheckbox;
+	private String groupsAddNewMembersInput;
+	private String groupsAddNewMembersAutocompleteOption1;
+	private String groupsAddNewMembersButton;
+	private String navigationSitebarNameId;
+	private String userOptions;
+	private String userOptionsLogout;
+
+	private String rightclickEditOption;
+	private String rightclickViewOption;
+	private String rightclickNewContentOption;
+	private String rightclickNewFolderOption;
+	private String rightclickDeleteOption;
+	private String rightclickChangeTemplateOption;
+	private String rightclickCutOption;
+	private String rightclickCopyOption;
+	private String rightclickDependenciesOption;
+	private String rightclickHistoryOption;
+	private String technologyLandingpage;
+	private String rightclickDuplicateOption;
+
+	private String articlesFolder;
+	private String articlesFolder2017;
+	private String articlesFolder1;
+	private String articlesFolderMenStylesForWinter;
+	private LinkedList<String> rightClickOptionsListInHomePage;
+	private LinkedList<String> rightClickOptionsListInCategoryLandingPage;
+	private LinkedList<String> rightClickOptionsListInMenStylesForWinterPage;
+	private String rightClickOptions;
+	private static Logger logger = LogManager
+			.getLogger(VerifyRightClickOptionsOfAPagesUnderPageStructureUsingDeveloperUser.class);
+
+	@BeforeMethod
+	public void beforeTest() {
+		userName = constantsPropertiesManager.getSharedExecutionConstants().getProperty("crafter.username");
+		password = constantsPropertiesManager.getSharedExecutionConstants().getProperty("crafter.password");
+		siteDropdownElementXPath = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("complexscenarios.general.sitedropdownmenuinnerxpath");
+		pagesTreeLink = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.sitecontent.expandpages");
+		pagesTree = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.expand_Pages_Tree");
+		homeContent = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.home_Content_Page");
+
+		rightclickEditOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.edit.option");
+		rightclickViewOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.view.option");
+		rightclickNewContentOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.Content.option");
+		rightclickNewFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.folder.option");
+		rightclickDeleteOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.delete.option");
+		rightclickChangeTemplateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.change.template.option");
+		rightclickCutOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.cut.option");
+		rightclickCopyOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.copy.option");
+		rightclickDependenciesOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.dependenciesoption");
+		rightclickHistoryOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.history.option");
+		technologyLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.pagestree.technology.landingpage");
+		rightclickDuplicateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.duplicate.option");
+		articlesFolder = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.articlesfolder");
+		articlesFolder2017 = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.articles.folder.2017");
+		articlesFolder1 = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.articles.folder.2017.1");
+		articlesFolderMenStylesForWinter = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.articles.folder.2017.1.menstylesforwinter");
+		newUserFirstNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.firstname");
+		newUserLastNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.lastname");
+		newUserEmailId = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("general.users.email");
+		newUserUserNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.username");
+		newUserPasswordId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.password");
+		newUserPasswordVerificationId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.passwordVerification");
+		newUserUserNameCreatedXpath = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.users.developerusernamecreated");
+		crafterLogo = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("users.crafterlogo");
+		expandPagesTree = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.expand_Pages_Tree");
+		adminConsole = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("general.adminconsole");
+		siteconfigGroupsOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("adminconsole.groups_option");
+		createSiteButton = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("home.createsitebutton");
+		addTouserIframe = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("complexscenarios.general.adduser.iframe");
+		editDeveloperGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.edit_developer_group_option");
+		groupsAddNewMembersCheckbox = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.add_new_members_checkbox");
+		groupsAddNewMembersInput = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.add_new_members_input");
+		groupsAddNewMembersAutocompleteOption1 = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.add_new_members_autocomplete_option1");
+		groupsAddNewMembersButton = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.add_new_members_button");
+		navigationSitebarNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("general.navigation_sitebar_name_id");
+		userOptions = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("dashboard.user_options");
+		userOptionsLogout = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.user_options_logout");
+		rightClickOptions = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.list.all.options");
+
+	}
+
+	public void rightClickHome() {
+		this.driverManager.waitUntilPageLoad();
+		this.driverManager.waitUntilSidebarOpens();
+
+		this.driverManager.waitUntilFolderOpens("xpath", pagesTree);
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homeContent);
+		dashboardPage.getDriverManager().contextClick("xpath", homeContent, false);
+
+	}
+
+	public void rightClickCategoryLandingPage() {
+		this.driverManager.waitUntilPageLoad();
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", technologyLandingpage);
+		dashboardPage.getDriverManager().contextClick("xpath", technologyLandingpage, false);
+		this.driverManager.waitForAnimation();
+	}
+
+	public void rightClickArticlesFolder1() {
+		this.driverManager.waitUntilPageLoad();
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1);
+		dashboardPage.getDriverManager().contextClick("xpath", articlesFolder1, false);
+		this.driverManager.waitForAnimation();
+	}
+
+	public void rightClickArticlesFolderMenStylesForWinter() {
+		this.driverManager.waitUntilPageLoad();
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+				articlesFolderMenStylesForWinter);
+		dashboardPage.getDriverManager().contextClick("xpath", articlesFolderMenStylesForWinter, false);
+		this.driverManager.waitForAnimation();
+	}
+
+	public void verifyEditOptionIsPresent(String section) {
+		WebElement rightclickEditOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickEditOption);
+		Assert.assertTrue(rightclickEditOptionElement.isDisplayed(),
+				"ERROR: Right click Edit Option is not present on right click of " + section);
+	}
+
+	public void verifyViewOptionIsPresent(String section) {
+		WebElement rightclickViewOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickViewOption);
+		Assert.assertTrue(rightclickViewOptionElement.isDisplayed(),
+				"ERROR: Right click View Option is not present on right click of " + section);
+	}
+
+	public void verifyNewContentOptionIsPresent(String section) {
+		WebElement rightclickNewContentOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewContentOption);
+		Assert.assertTrue(rightclickNewContentOptionElement.isDisplayed(),
+				"ERROR: Right click New Content Option is not present on right click of " + section);
+	}
+
+	public void verifyNewFolderOptionIsPresent(String section) {
+		WebElement rightclickNewFolderOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewFolderOption);
+		Assert.assertTrue(rightclickNewFolderOptionElement.isDisplayed(),
+				"ERROR: Right click New Folder Option is not present on right click of " + section);
+	}
+
+	public void verifyDeleteOptionIsPresent(String section) {
+		WebElement rightclickDeleteOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDeleteOption);
+		Assert.assertTrue(rightclickDeleteOptionElement.isDisplayed(),
+				"ERROR: Right click Delete Option is not present on right click of " + section);
+	}
+
+	public void verifyChangeTemplateOptionIsPresent(String section) {
+		WebElement rightclickChangeTemplateOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickChangeTemplateOption);
+		Assert.assertTrue(rightclickChangeTemplateOptionElement.isDisplayed(),
+				"ERROR: Right click Change Template Option is not present on right click of " + section);
+	}
+
+	public void verifyCutOptionIsPresent(String section) {
+		WebElement rightclickCutOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickCutOption);
+		Assert.assertTrue(rightclickCutOptionElement.isDisplayed(),
+				"ERROR: Right click Cut Option is not present on right click of " + section);
+	}
+
+	public void verifyCopyOptionIsPresent(String section) {
+		WebElement rightclickCopyOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickCopyOption);
+		Assert.assertTrue(rightclickCopyOptionElement.isDisplayed(),
+				"ERROR: Right click Copy Option is not present on right click of " + section);
+	}
+
+	public void verifyDependenciesOptionIsPresent(String section) {
+		WebElement rightclickDependenciesOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDependenciesOption);
+		Assert.assertTrue(rightclickDependenciesOptionElement.isDisplayed(),
+				"ERROR: Right click Dependencies Option is not present on right click of " + section);
+	}
+
+	public void verifyHistoryOptionIsPresent(String section) {
+		WebElement rightclickHistoryOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickHistoryOption);
+		Assert.assertTrue(rightclickHistoryOptionElement.isDisplayed(),
+				"ERROR: Right click History Option is not present on right click of " + section);
+	}
+
+	public void verifyDuplicateOptionIsPresent(String section) {
+		WebElement rightclickDuplicateOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDuplicateOption);
+		Assert.assertTrue(rightclickDuplicateOptionElement.isDisplayed(),
+				"ERROR: Right click Duplicate Option is not present on right click of " + section);
+	}
+
+	public void step4() {
+		// Step 4 Right Right click on "Home" and verify options
+		this.rightClickHome();
+
+		driverManager.usingContextMenu(() -> {
+			
+			logger.info("Checking that only the expected options are listed");
+			rightClickOptionsListInHomePage = new LinkedList<String>();
+			rightClickOptionsListInHomePage.add(0, "Edit");
+			rightClickOptionsListInHomePage.add(1, "View");
+			rightClickOptionsListInHomePage.add(2, "New Content");
+			rightClickOptionsListInHomePage.add(3, "New Folder");
+			rightClickOptionsListInHomePage.add(4, "Change Template");
+			rightClickOptionsListInHomePage.add(5, "Cut");
+			rightClickOptionsListInHomePage.add(6, "Copy");
+			rightClickOptionsListInHomePage.add(7, "Dependencies");
+			rightClickOptionsListInHomePage.add(8, "History");
+
+			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
+					.findElements(By.xpath(rightClickOptions));
+			int currentIndex = 0;
+			for (WebElement element : rightClickOptionsList) {
+				this.driverManager.waitForAnimation();
+				this.driverManager.waitUntilSidebarOpens();
+				Assert.assertTrue(element.getText().equals(rightClickOptionsListInHomePage.get(currentIndex)),
+						"ERROR: Link Option: " + element.getText()
+								+ " is not in the correct order in the HomePage, check that the correct options are listed");
+				currentIndex++;
+			}
+
+			String section = "Step 4 Right Click on 'Home'";
+
+			verifyEditOptionIsPresent(section);
+			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
+			verifyCopyOptionIsPresent(section);
+			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
+
+			this.driverManager.getDriver().navigate().refresh();
+			this.driverManager.waitForAnimation();
+		});
+	}
+
+	public void step6() {
+		// Step 6 Right click on any Category Landing page and verify options
+		this.rightClickCategoryLandingPage();
+
+		driverManager.usingContextMenu(() -> {
+			
+			logger.info("Checking that only the expected options are listed");
+			rightClickOptionsListInCategoryLandingPage = new LinkedList<String>();
+			rightClickOptionsListInCategoryLandingPage.add(0, "Edit");
+			rightClickOptionsListInCategoryLandingPage.add(1, "View");
+			rightClickOptionsListInCategoryLandingPage.add(2, "New Content");
+			rightClickOptionsListInCategoryLandingPage.add(3, "New Folder");
+			rightClickOptionsListInCategoryLandingPage.add(4, "Delete");
+			rightClickOptionsListInCategoryLandingPage.add(5, "Change Template");
+			rightClickOptionsListInCategoryLandingPage.add(6, "Cut");
+			rightClickOptionsListInCategoryLandingPage.add(7, "Copy");
+			rightClickOptionsListInCategoryLandingPage.add(8, "Duplicate");
+			rightClickOptionsListInCategoryLandingPage.add(9, "Dependencies");
+			rightClickOptionsListInCategoryLandingPage.add(10, "History");
+
+			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
+					.findElements(By.xpath(rightClickOptions));
+			int currentIndex = 0;
+			for (WebElement element : rightClickOptionsList) {
+				this.driverManager.waitForAnimation();
+				this.driverManager.waitUntilSidebarOpens();
+				Assert.assertTrue(
+						element.getText().equals(rightClickOptionsListInCategoryLandingPage.get(currentIndex)),
+						"ERROR: Link Option: " + element.getText()
+								+ " is not in the correct order in the selected category landing page, check that the correct options are listed");
+				currentIndex++;
+			}
+
+			String section = "Step 6 Right click on a 'Category Landing' page";
+
+			verifyEditOptionIsPresent(section);
+			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyDeleteOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
+			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
+			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
+
+			this.driverManager.getDriver().navigate().refresh();
+			this.driverManager.waitForAnimation();
+		});
+	}
+
+	public void step8() {
+		// Step 8 Click on the + of folder 2017
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder2017);
+
+		this.driverManager.waitUntilContentTooltipIsHidden();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder2017).click();
+		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder2017);
+	}
+
+
+	public void step10() {
+		// Step 11 Right click on any of the article (Men Styles For Winter)
+		this.driverManager.waitForAnimation();
+
+		this.rightClickArticlesFolderMenStylesForWinter();
+
+		driverManager.usingContextMenu(() -> {
+			
+			logger.info("Checking that only the expected options are listed");
+			rightClickOptionsListInMenStylesForWinterPage = new LinkedList<String>();
+			rightClickOptionsListInMenStylesForWinterPage.add(0, "Edit");
+			rightClickOptionsListInMenStylesForWinterPage.add(1, "View");
+			rightClickOptionsListInMenStylesForWinterPage.add(2, "New Content");
+			rightClickOptionsListInMenStylesForWinterPage.add(3, "New Folder");
+			rightClickOptionsListInMenStylesForWinterPage.add(4, "Delete");
+			rightClickOptionsListInMenStylesForWinterPage.add(5, "Change Template");
+			rightClickOptionsListInMenStylesForWinterPage.add(6, "Cut");
+			rightClickOptionsListInMenStylesForWinterPage.add(7, "Copy");
+			rightClickOptionsListInMenStylesForWinterPage.add(8, "Duplicate");
+			rightClickOptionsListInMenStylesForWinterPage.add(9, "Dependencies");
+			rightClickOptionsListInMenStylesForWinterPage.add(10, "History");
+
+			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
+					.findElements(By.xpath(rightClickOptions));
+			int currentIndex = 0;
+			for (WebElement element : rightClickOptionsList) {
+				this.driverManager.waitForAnimation();
+				this.driverManager.waitUntilSidebarOpens();
+				Assert.assertTrue(
+						element.getText().equals(rightClickOptionsListInMenStylesForWinterPage.get(currentIndex)),
+						"ERROR: Link Option: " + element.getText()
+								+ " is not in the correct order in the Men Styles For Witner page, check that the correct options are listed");
+				currentIndex++;
+			}
+
+			String section = "Step 11 Right click on folder articles -2017- 1- Men Styles For Winter;";
+
+			verifyEditOptionIsPresent(section);
+			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyDeleteOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
+			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
+			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
+
+			this.driverManager.getDriver().navigate().refresh();
+			this.driverManager.waitForAnimation();
+		});
+		
+	}
+	
+	public void login(String user, String loginpassword) {
+
+		// login to application
+		loginPage.loginToCrafter(user, loginpassword);
+
+		// Wait for login page to close
+		driverManager.waitUntilLoginCloses();
+	}
+	
+	public void addNewUser() {
+
+		// click On Users option
+
+		createSitePage.clickOnUsersOption();
+
+		// click on new user button
+
+		usersPage.clickOnNewUser();
+
+		// Follow the form
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("developer");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserLastNameId)
+				.sendKeys("Last Name");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserEmailId)
+				.sendKeys("developer@email.com");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("developer");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("developer");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordVerificationId)
+				.sendKeys("developer");
+
+		// Save Button
+		usersPage.clickOnSaveNewUser();
+
+		// Assert new users created is present
+		WebElement newUserCreated = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				newUserUserNameCreatedXpath);
+
+		Assert.assertTrue(newUserCreated.isDisplayed(), "ERROR: Recently created user is not displayed");
+
+		// Switch to the form
+
+		driverManager.getDriver().navigate().refresh();
+
+		driverManager.getDriver().switchTo().defaultContent();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+
+				crafterLogo);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+
+				crafterLogo).click();
+	}
+	
+	private void goToSiteContentPagesStructure() {
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+				createSiteButton);
+		
+		this.driverManager.waitForAnimation();
+		homePage.goToPreviewPage();
+
+		if (this.driverManager.isElementPresentByXpath(siteDropdownElementXPath))
+			this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", siteDropdownElementXPath).click();
+		else
+			throw new NoSuchElementException(
+					"Site creation process is taking too long time and the element was not found");
+	}
+	
+	public void addUserToAuthorGroup() {
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", adminConsole);
+
+		WebElement siteConfigButton = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				adminConsole);
+
+		siteConfigButton.click();
+
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", siteconfigGroupsOption);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", siteconfigGroupsOption)
+
+				.click();
+
+		driverManager.getDriver().switchTo().defaultContent();
+
+		this.driverManager.getDriver().switchTo()
+
+				.frame(this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", addTouserIframe));
+
+		this.driverManager.isElementPresentAndClickableByXpath(addTouserIframe);
+
+		this.driverManager.getDriver().switchTo().activeElement();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+
+				editDeveloperGroupOption);
+
+		this.driverManager
+
+				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editDeveloperGroupOption)
+
+				.click();
+
+		driverManager.getDriver().switchTo().defaultContent();
+
+		this.driverManager.getDriver().switchTo()
+
+				.frame(this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", addTouserIframe));
+
+		this.driverManager.isElementPresentAndClickableByXpath(addTouserIframe);
+
+		this.driverManager.getDriver().switchTo().activeElement();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersCheckbox);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersCheckbox)
+
+				.click();
+
+		this.driverManager
+
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersInput)
+
+				.sendKeys("developer");
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+
+				groupsAddNewMembersAutocompleteOption1);
+
+		this.driverManager
+
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+
+						groupsAddNewMembersAutocompleteOption1)
+
+				.click();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+
+				groupsAddNewMembersButton);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
+
+				groupsAddNewMembersButton).click();
+
+		driverManager.getDriver().switchTo().defaultContent();
+
+		this.driverManager.getDriver().switchTo().activeElement();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", navigationSitebarNameId);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", navigationSitebarNameId)
+				.click();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", adminConsole);
+	}
+	
+	private void logoutFromCrafter() {
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", userOptions);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", userOptions).click();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", userOptionsLogout);
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", userOptionsLogout).click();
+
+	}
+
+	@Test(priority = 0)
+	public void verifyRightClickOptionsOfAPagesUnderPageStructureUsingDeveloperUser() {
+		
+		this.login(userName, password);
+
+		logger.info("Adding New User");
+
+		this.addNewUser();
+
+		this.driverManager.getDriver().navigate().refresh();
+
+		logger.info("Go to Site Preview");
+
+		this.goToSiteContentPagesStructure();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", expandPagesTree);
+
+		this.driverManager.waitUntilSidebarOpens();
+
+		logger.info("Add previous created user to Developer Group");
+
+		this.addUserToAuthorGroup();
+
+		// logout from Crafter
+		logger.info("logout from Crafter");
+		this.logoutFromCrafter();
+
+		// login to application with developer user
+		logger.info("login to application with developer user");
+		loginPage.loginToCrafter("developer", "developer");
+
+		logger.info("Go to Preview Page");
+		this.homePage.goToPreviewPage();
+
+		this.driverManager.waitForAnimation();
+
+		// Expand the site bar
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", siteDropdownElementXPath);
+		
+		Assert.assertTrue(this.driverManager.isElementPresentAndClickableByXpath(siteDropdownElementXPath));
+		
+		// Step 2 Expand the site bar Step No needed
+		// Step 3 Click on Pages tree
+		WebElement pagesTreeLinkElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				pagesTreeLink);
+		pagesTreeLinkElement.click();
+		this.driverManager.waitUntilFolderOpens("xpath", pagesTreeLink);
+
+		// Step 4 Right Right click on "Home" and verify options
+		this.step4();
+
+		// Step 5 Click on the + of Home tree
+		this.dashboardPage.expandHomeTree();
+
+		// Step 6 Right click on any Category Landing page and verify options
+		this.step6();
+
+		// Step 7 Click on the + of Articles tree
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder);
+
+		this.driverManager.waitUntilContentTooltipIsHidden();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder).click();
+		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder);
+
+		// Step 8 Click on the + of folder 2017
+		this.step8();
+
+		// Step 9 Click on the + of folder "1"
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1).click();
+		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder1);
+
+		// Step 10 Right click on any of the article (Men Styles For Winter)
+		this.step10();
+	}
+}
