+++ src/test/java/org/craftercms/studio/test/cases/sitedropdowntestcases/VerifyRightClickOptionsOfAPagesUnderPageStructureUsingReviewerUser.java
@@ -20,8 +20,8 @@ import org.openqa.selenium.WebElement;
  * @author Juan Camacho A
  *
  */
-// Test Case ID:106
-public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThePagesStructureIsEditedUsingPublisherUser
+//Test Case Studio- Site Dropdown ID:13
+public class VerifyRightClickOptionsOfAPagesUnderPageStructureUsingReviewerUser
 		extends BaseTest {
 
 	private String userName;
@@ -30,12 +30,6 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private String pagesTreeLink;
 	private String pagesTree;
 	private String homeContent;
-	private String createFormFrameElementCss;
-	private String homePageEditedName;
-	private String homePageEditedExpand;
-	private String menStylesEditedPageName;
-	private String menStylesForWinterEditedPageName;
-	private String rightClickOptions;
 	
 	private String newUserFirstNameId;
 	private String newUserLastNameId;
@@ -50,7 +44,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private String adminConsole;
 	private String siteconfigGroupsOption;
 	private String addTouserIframe;
-	private String editPublisherGroupOption;
+	private String editReviewerGroupOption;
 	private String groupsAddNewMembersCheckbox;
 	private String groupsAddNewMembersInput;
 	private String groupsAddNewMembersAutocompleteOption1;
@@ -59,18 +53,12 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private String userOptions;
 	private String userOptionsLogout;
 
-	private String rightclickEditOption;
 	private String rightclickViewOption;
-	private String rightclickNewContentOption;
-	private String rightclickNewFolderOption;
-	private String rightclickDeleteOption;
-	private String rightclickCutOption;
 	private String rightclickCopyOption;
-	private String rightclickDuplicateOption;
-	private String rightclickApproveAndPublishOption;
 	private String rightclickDependenciesOption;
-	private String rightclickHistoryOption;
-	private String styleLandingpage;
+	private String technologyLandingpage;
+	private String rightclickDuplicateOption;
+	private String rightclickRenameFolderOption;
 	private String articlesFolder;
 	private String articlesFolder2017;
 	private String articlesFolder1;
@@ -78,8 +66,9 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private LinkedList<String> rightClickOptionsListInHomePage;
 	private LinkedList<String> rightClickOptionsListInCategoryLandingPage;
 	private LinkedList<String> rightClickOptionsListInMenStylesForWinterPage;
-	private static Logger logger = LogManager.getLogger(
-			VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThePagesStructureIsEditedUsingPublisherUser.class);
+	private String rightClickOptions;
+	private static Logger logger = LogManager
+			.getLogger(VerifyRightClickOptionsOfAPagesUnderPageStructureUsingReviewerUser.class);
 
 	@BeforeMethod
 	public void beforeTest() {
@@ -93,39 +82,19 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				.getProperty("dashboard.expand_Pages_Tree");
 		homeContent = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.home_Content_Page");
-		homePageEditedName = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.new_home_Content_Page_name_edited");
-		menStylesEditedPageName = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.pagestree.style.landingpage.edited");
-		homePageEditedExpand = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.new_home_Content_Page_name_edited_expand");
-		menStylesForWinterEditedPageName = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.pagestree.articlesfolder.menstylesforwinter.page.edited");
-		
-		rightclickEditOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.edit.option");
+
 		rightclickViewOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.view.option");
-		rightclickNewContentOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.new.Content.option");
-		rightclickNewFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.new.folder.option");
-		rightclickDeleteOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.delete.option");
-		rightclickCutOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.cut.option");
 		rightclickCopyOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.copy.option");
-		rightclickApproveAndPublishOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.approveandpublish.option");
 		rightclickDependenciesOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.dependenciesoption");
-		rightclickHistoryOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.history.option");
-		styleLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.pagestree.style.landingpage");
+		technologyLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.pagestree.technology.landingpage");
 		rightclickDuplicateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.duplicate.option");
+		rightclickRenameFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.rename.folder.option");
 		articlesFolder = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.articlesfolder");
 		articlesFolder2017 = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -134,10 +103,6 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				.getProperty("dashboard.articles.folder.2017.1");
 		articlesFolderMenStylesForWinter = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.articles.folder.2017.1.menstylesforwinter");
-		createFormFrameElementCss = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("complexscenarios.general.createformframe");
-		rightClickOptions = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.list.all.options");
 		newUserFirstNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("general.users.firstname");
 		newUserLastNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -150,7 +115,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		newUserPasswordVerificationId = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("general.users.passwordVerification");
 		newUserUserNameCreatedXpath = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("general.users.publisherusernamecreated");
+				.getProperty("general.users.reviewerusernamecreated");
 		crafterLogo = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("users.crafterlogo");
 		expandPagesTree = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.expand_Pages_Tree");
@@ -161,8 +126,8 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				.getProperty("home.createsitebutton");
 		addTouserIframe = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("complexscenarios.general.adduser.iframe");
-		editPublisherGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("groups.edit_publisher_group_option");
+		editReviewerGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.edit_reviewer_group_option");
 		groupsAddNewMembersCheckbox = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("groups.add_new_members_checkbox");
 		groupsAddNewMembersInput = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -176,18 +141,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		userOptions = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("dashboard.user_options");
 		userOptionsLogout = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.user_options_logout");
+		rightClickOptions = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.list.all.options");
 
 	}
 
-	public void rightClickHomePage() {
+	public void rightClickHome() {
 		this.driverManager.waitUntilPageLoad();
 		this.driverManager.waitUntilSidebarOpens();
 
 		this.driverManager.waitUntilFolderOpens("xpath", pagesTree);
 		this.driverManager.waitForAnimation();
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homePageEditedName);
-		dashboardPage.getDriverManager().contextClick("xpath", homePageEditedName, false);
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homeContent);
+		dashboardPage.getDriverManager().contextClick("xpath", homeContent, false);
 
 	}
 
@@ -196,8 +163,18 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.waitUntilSidebarOpens();
 		this.driverManager.waitForAnimation();
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", menStylesEditedPageName);
-		dashboardPage.getDriverManager().contextClick("xpath", menStylesEditedPageName, false);
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", technologyLandingpage);
+		dashboardPage.getDriverManager().contextClick("xpath", technologyLandingpage, false);
+		this.driverManager.waitForAnimation();
+	}
+
+	public void rightClickArticlesFolder1() {
+		this.driverManager.waitUntilPageLoad();
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.waitForAnimation();
+
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1);
+		dashboardPage.getDriverManager().contextClick("xpath", articlesFolder1, false);
 		this.driverManager.waitForAnimation();
 	}
 
@@ -207,16 +184,9 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.waitForAnimation();
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
-				menStylesForWinterEditedPageName);
-		dashboardPage.getDriverManager().contextClick("xpath", menStylesForWinterEditedPageName, false);
+				articlesFolderMenStylesForWinter);
+		dashboardPage.getDriverManager().contextClick("xpath", articlesFolderMenStylesForWinter, false);
 		this.driverManager.waitForAnimation();
-	}
-
-	public void verifyEditOptionIsPresent(String section) {
-		WebElement rightclickEditOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
-				rightclickEditOption);
-		Assert.assertTrue(rightclickEditOptionElement.isDisplayed(),
-				"ERROR: Right click Edit Option is not present on right click of " + section);
 	}
 
 	public void verifyViewOptionIsPresent(String section) {
@@ -226,46 +196,11 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				"ERROR: Right click View Option is not present on right click of " + section);
 	}
 
-	public void verifyNewContentOptionIsPresent(String section) {
-		WebElement rightclickNewContentOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewContentOption);
-		Assert.assertTrue(rightclickNewContentOptionElement.isDisplayed(),
-				"ERROR: Right click New Content Option is not present on right click of " + section);
-	}
-
-	public void verifyNewFolderOptionIsPresent(String section) {
-		WebElement rightclickNewFolderOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewFolderOption);
-		Assert.assertTrue(rightclickNewFolderOptionElement.isDisplayed(),
-				"ERROR: Right click New Folder Option is not present on right click of " + section);
-	}
-
-	public void verifyDeleteOptionIsPresent(String section) {
-		WebElement rightclickDeleteOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDeleteOption);
-		Assert.assertTrue(rightclickDeleteOptionElement.isDisplayed(),
-				"ERROR: Right click Delete Option is not present on right click of " + section);
-	}
-
-	public void verifyCutOptionIsPresent(String section) {
-		WebElement rightclickCutOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
-				rightclickCutOption);
-		Assert.assertTrue(rightclickCutOptionElement.isDisplayed(),
-				"ERROR: Right click Cut Option is not present on right click of " + section);
-	}
-
 	public void verifyCopyOptionIsPresent(String section) {
 		WebElement rightclickCopyOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 				rightclickCopyOption);
 		Assert.assertTrue(rightclickCopyOptionElement.isDisplayed(),
 				"ERROR: Right click Copy Option is not present on right click of " + section);
-	}
-	
-	public void verifyApproveAndPublishOptionIsPresent(String section) {
-		WebElement rightclickApproveAndPublishOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickApproveAndPublishOption);
-		Assert.assertTrue(rightclickApproveAndPublishOptionElement.isDisplayed(),
-				"ERROR: Right click Approve and Publish Option is not present on right click of " + section);
 	}
 
 	public void verifyDependenciesOptionIsPresent(String section) {
@@ -275,13 +210,6 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				"ERROR: Right click Dependencies Option is not present on right click of " + section);
 	}
 
-	public void verifyHistoryOptionIsPresent(String section) {
-		WebElement rightclickHistoryOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickHistoryOption);
-		Assert.assertTrue(rightclickHistoryOptionElement.isDisplayed(),
-				"ERROR: Right click History Option is not present on right click of " + section);
-	}
-
 	public void verifyDuplicateOptionIsPresent(String section) {
 		WebElement rightclickDuplicateOptionElement = this.driverManager
 				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDuplicateOption);
@@ -289,22 +217,24 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				"ERROR: Right click Duplicate Option is not present on right click of " + section);
 	}
 
+	public void verifyRenameFolderOptionIsPresent(String section) {
+		WebElement rightclickRenameFolderOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickRenameFolderOption);
+		Assert.assertTrue(rightclickRenameFolderOptionElement.isDisplayed(),
+				"ERROR: Right click Rename Folder Option is not present on right click of " + section);
+	}
+
 	public void step4() {
 		// Step 4 Right Right click on "Home" and verify options
-		this.rightClickHomePage();
+		this.rightClickHome();
 
 		driverManager.usingContextMenu(() -> {
+			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInHomePage = new LinkedList<String>();
-			rightClickOptionsListInHomePage.add(0, "Edit");
-			rightClickOptionsListInHomePage.add(1, "View");
-			rightClickOptionsListInHomePage.add(2, "New Content");
-			rightClickOptionsListInHomePage.add(3, "New Folder");
-			rightClickOptionsListInHomePage.add(4, "Cut");
-			rightClickOptionsListInHomePage.add(5, "Copy");
-			rightClickOptionsListInHomePage.add(6, "Approve & Publish");
-			rightClickOptionsListInHomePage.add(7, "Dependencies");
-			rightClickOptionsListInHomePage.add(8, "History");
+			rightClickOptionsListInHomePage.add(0, "View");
+			rightClickOptionsListInHomePage.add(1, "Copy");
+			rightClickOptionsListInHomePage.add(2, "Dependencies");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -319,41 +249,27 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 			}
 
 			String section = "Step 4 Right Click on 'Home'";
-
-			verifyEditOptionIsPresent(section);
+			
 			verifyViewOptionIsPresent(section);
-			verifyNewContentOptionIsPresent(section);
-			verifyNewFolderOptionIsPresent(section);
-			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
-			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
-			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
-			this.driverManager.waitUntilSidebarOpens();
 		});
 	}
 
 	public void step6() {
 		// Step 6 Right click on any Category Landing page and verify options
 		this.rightClickCategoryLandingPage();
-		
+
 		driverManager.usingContextMenu(() -> {
+			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInCategoryLandingPage = new LinkedList<String>();
-			rightClickOptionsListInCategoryLandingPage.add(0, "Edit");
-			rightClickOptionsListInCategoryLandingPage.add(1, "View");
-			rightClickOptionsListInCategoryLandingPage.add(2, "New Content");
-			rightClickOptionsListInCategoryLandingPage.add(3, "New Folder");
-			rightClickOptionsListInCategoryLandingPage.add(4, "Delete");
-			rightClickOptionsListInCategoryLandingPage.add(5, "Cut");
-			rightClickOptionsListInCategoryLandingPage.add(6, "Copy");
-			rightClickOptionsListInCategoryLandingPage.add(7, "Duplicate");
-			rightClickOptionsListInCategoryLandingPage.add(8, "Approve & Publish");
-			rightClickOptionsListInCategoryLandingPage.add(9, "Dependencies");
-			rightClickOptionsListInCategoryLandingPage.add(10, "History");
+			rightClickOptionsListInCategoryLandingPage.add(0, "View");
+			rightClickOptionsListInCategoryLandingPage.add(1, "Copy");
+			rightClickOptionsListInCategoryLandingPage.add(2, "Dependencies");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -369,18 +285,9 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 			}
 
 			String section = "Step 6 Right click on a 'Category Landing' page";
-
-			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
-			verifyNewContentOptionIsPresent(section);
-			verifyNewFolderOptionIsPresent(section);
-			verifyDeleteOptionIsPresent(section);
-			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
-			verifyDuplicateOptionIsPresent(section);
-			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
-			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
@@ -397,8 +304,9 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder2017);
 	}
 
+
 	public void step10() {
-		// Step 10 Right click on any of the article (Men Styles For Winter)
+		// Step 11 Right click on any of the article (Men Styles For Winter)
 		this.driverManager.waitForAnimation();
 
 		this.rightClickArticlesFolderMenStylesForWinter();
@@ -407,17 +315,9 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInMenStylesForWinterPage = new LinkedList<String>();
-			rightClickOptionsListInMenStylesForWinterPage.add(0, "Edit");
-			rightClickOptionsListInMenStylesForWinterPage.add(1, "View");
-			rightClickOptionsListInMenStylesForWinterPage.add(2, "New Content");
-			rightClickOptionsListInMenStylesForWinterPage.add(3, "New Folder");
-			rightClickOptionsListInMenStylesForWinterPage.add(4, "Delete");
-			rightClickOptionsListInMenStylesForWinterPage.add(5, "Cut");
-			rightClickOptionsListInMenStylesForWinterPage.add(6, "Copy");
-			rightClickOptionsListInMenStylesForWinterPage.add(7, "Duplicate");
-			rightClickOptionsListInMenStylesForWinterPage.add(8, "Approve & Publish");
-			rightClickOptionsListInMenStylesForWinterPage.add(9, "Dependencies");
-			rightClickOptionsListInMenStylesForWinterPage.add(10, "History");
+			rightClickOptionsListInMenStylesForWinterPage.add(0, "View");
+			rightClickOptionsListInMenStylesForWinterPage.add(1, "Copy");
+			rightClickOptionsListInMenStylesForWinterPage.add(2, "Dependencies");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -428,102 +328,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				Assert.assertTrue(
 						element.getText().equals(rightClickOptionsListInMenStylesForWinterPage.get(currentIndex)),
 						"ERROR: Link Option: " + element.getText()
-								+ " is not in the correct order in the Men Styles For Winter page, check that the correct options are listed");
+								+ " is not in the correct order in the Men Styles For Witner page, check that the correct options are listed");
 				currentIndex++;
 			}
 
 			String section = "Step 11 Right click on folder articles -2017- 1- Men Styles For Winter;";
 
-			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
-			verifyNewContentOptionIsPresent(section);
-			verifyNewFolderOptionIsPresent(section);
-			verifyDeleteOptionIsPresent(section);
-			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
-			verifyDuplicateOptionIsPresent(section);
-			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
-			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
 		});
-
-	}
-
-	public void editHomePage() {
-		// Step 3 Edit the Home Page and Save
-
-		dashboardPage.rightClickEditOnAPresentPage(homeContent);
-
-		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
-
-			// creating random values for URL field and InternalName field
-
-			String randomInternalName = "HomePageEdited";
-
-			// Set basics fields of the new content created
-
-			dashboardPage.editInternalName(randomInternalName);
-		});
-
-		this.driverManager.waitUntilSidebarOpens();
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", homePageEditedName);
-
-	}
-
-	public void editLandingPageStylePage() {
-
-		dashboardPage.rightClickEditOnAPresentPage(styleLandingpage);
-
-		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
-
-			// creating random values for URL field and InternalName field
-
-			String randomInternalName = "StylePageEdited";
-
-			// Set basics fields of the new content created
-
-			dashboardPage.editInternalName(randomInternalName);
-		});
-
-		this.driverManager.waitUntilSidebarOpens();
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", menStylesEditedPageName);
-
-	}
-
-	public void editMenStylesForWinterPage() {
-
-		dashboardPage.rightClickEditOnAPresentPage(articlesFolderMenStylesForWinter);
-
-		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
-
-			// creating random values for URL field and InternalName field
-
-			String randomInternalName = "MenStylesForWinterEdited";
-
-			// Set basics fields of the new content created
-
-			dashboardPage.editInternalName(randomInternalName);
-		});
-
-		this.driverManager.waitUntilSidebarOpens();
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", menStylesForWinterEditedPageName);
-
-	}
-
-	public void expandHomeTree() {
-		this.driverManager.waitUntilSidebarOpens();
-		this.driverManager.waitForAnimation();
-
-		// Verify if the home tree is already expanded
-		if (!(this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", homePageEditedName)
-				.getAttribute("class").contains("open"))) {
-			this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homePageEditedExpand)
-					.click();
-		}
-
+		
 	}
 	
 	public void login(String user, String loginpassword) {
@@ -546,20 +364,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		usersPage.clickOnNewUser();
 
 		// Follow the form
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("publisher");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("reviewer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserLastNameId)
 				.sendKeys("Last Name");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserEmailId)
-				.sendKeys("email@email.com");
+				.sendKeys("reviewer@email.com");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("publisher");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("reviewer");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("publisher");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("reviewer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordVerificationId)
-				.sendKeys("publisher");
+				.sendKeys("reviewer");
 
 		// Save Button
 		usersPage.clickOnSaveNewUser();
@@ -600,7 +418,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 					"Site creation process is taking too long time and the element was not found");
 	}
 	
-	public void addUserToPublisherGroup() {
+	public void addUserToAuthorGroup() {
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", adminConsole);
 
@@ -629,11 +447,11 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
 
-				editPublisherGroupOption);
+				editReviewerGroupOption);
 
 		this.driverManager
 
-				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editPublisherGroupOption)
+				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editReviewerGroupOption)
 
 				.click();
 
@@ -657,7 +475,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 				.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersInput)
 
-				.sendKeys("publisher");
+				.sendKeys("reviewer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 
@@ -704,8 +522,8 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	}
 
 	@Test(priority = 0)
-	public void verifyProperOptionDisplayedOnPagesStructureWithPublisherUser() {
-
+	public void verifyRightClickOptionsOfAPagesUnderPageStructureUsingReviewerUser() {
+		
 		this.login(userName, password);
 
 		logger.info("Adding New User");
@@ -722,17 +540,17 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 		this.driverManager.waitUntilSidebarOpens();
 
-		logger.info("Add previous created user to Publisher Group");
+		logger.info("Add previous created user to Reviewer Group");
 
-		this.addUserToPublisherGroup();
+		this.addUserToAuthorGroup();
 
 		// logout from Crafter
 		logger.info("logout from Crafter");
 		this.logoutFromCrafter();
 
-		// login to application with publisher user
-		logger.info("login to application with publisher user");
-		loginPage.loginToCrafter("publisher", "publisher");
+		// login to application with reviewer user
+		logger.info("login to application with reviewer user");
+		loginPage.loginToCrafter("reviewer", "reviewer");
 
 		logger.info("Go to Preview Page");
 		this.homePage.goToPreviewPage();
@@ -751,20 +569,16 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		pagesTreeLinkElement.click();
 		this.driverManager.waitUntilFolderOpens("xpath", pagesTreeLink);
 
-		this.editHomePage();
-
-		logger.info("Step 4 Right Right click on Home and verify options");
+		// Step 4 Right Right click on "Home" and verify options
 		this.step4();
 
-		logger.info("Step 5 Click on the + of Home tree and Edit Home Page");
-		this.expandHomeTree();
+		// Step 5 Click on the + of Home tree
+		this.dashboardPage.expandHomeTree();
 
-		logger.info("Step 6 Right click on any Category Landing page style and verify options");
-		this.editLandingPageStylePage();
-
+		// Step 6 Right click on any Category Landing page and verify options
 		this.step6();
 
-		logger.info("Step 7 Click on the + of Articles tree");
+		// Step 7 Click on the + of Articles tree
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder);
 
 		this.driverManager.waitUntilContentTooltipIsHidden();
@@ -772,17 +586,14 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder);
 
-		logger.info("Step 8 Click on the + of folder 2017");
+		// Step 8 Click on the + of folder 2017
 		this.step8();
 
-		logger.info(
-				"Step 9 Click on the + of folder 1 and Edit then article page Men Styles For Winter under articles/2017/1");
+		// Step 9 Click on the + of folder "1"
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder1);
 
-		this.editMenStylesForWinterPage();
-
-		logger.info("Step 10 Right click on any of the article Men Styles For Winter");
+		// Step 10 Right click on any of the article (Men Styles For Winter)
 		this.step10();
 	}
 }
