+++ src/test/java/org/craftercms/studio/test/cases/sitedropdowntestcases/VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingAuthorUser.java
@@ -20,8 +20,8 @@ import org.openqa.selenium.WebElement;
  * @author Juan Camacho A
  *
  */
-// Test Case ID:101
-public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClickIsPerformedOnAnyElementOfThePagesStructureWithReviewerUser
+//Test Case Studio- Site Dropdown ID:16
+public class VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingAuthorUser
 		extends BaseTest {
 
 	private String userName;
@@ -30,6 +30,12 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 	private String pagesTreeLink;
 	private String pagesTree;
 	private String homeContent;
+	private String createFormFrameElementCss;
+	private String homePageEditedName;
+	private String homePageEditedExpand;
+	private String menStylesEditedPageName;
+	private String menStylesForWinterEditedPageName;
+	private String rightClickOptions;
 	
 	private String newUserFirstNameId;
 	private String newUserLastNameId;
@@ -44,7 +50,7 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 	private String adminConsole;
 	private String siteconfigGroupsOption;
 	private String addTouserIframe;
-	private String editReviewerGroupOption;
+	private String editAuthorGroupOption;
 	private String groupsAddNewMembersCheckbox;
 	private String groupsAddNewMembersInput;
 	private String groupsAddNewMembersAutocompleteOption1;
@@ -53,12 +59,17 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 	private String userOptions;
 	private String userOptionsLogout;
 
+	private String rightclickEditOption;
 	private String rightclickViewOption;
+	private String rightclickNewContentOption;
+	private String rightclickNewFolderOption;
+	private String rightclickCutOption;
 	private String rightclickCopyOption;
+	private String rightclickRequestPublishOption;
 	private String rightclickDependenciesOption;
-	private String technologyLandingpage;
+	private String rightclickHistoryOption;
+	private String styleLandingpage;
 	private String rightclickDuplicateOption;
-	private String rightclickRenameFolderOption;
 	private String articlesFolder;
 	private String articlesFolder2017;
 	private String articlesFolder1;
@@ -66,9 +77,8 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 	private LinkedList<String> rightClickOptionsListInHomePage;
 	private LinkedList<String> rightClickOptionsListInCategoryLandingPage;
 	private LinkedList<String> rightClickOptionsListInMenStylesForWinterPage;
-	private String rightClickOptions;
-	private static Logger logger = LogManager
-			.getLogger(VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClickIsPerformedOnAnyElementOfThePagesStructureWithReviewerUser.class);
+	private static Logger logger = LogManager.getLogger(
+			VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingAuthorUser.class);
 
 	@BeforeMethod
 	public void beforeTest() {
@@ -82,19 +92,38 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				.getProperty("dashboard.expand_Pages_Tree");
 		homeContent = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.home_Content_Page");
+		homePageEditedName = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.new_home_Content_Page_name_edited");
+		menStylesEditedPageName = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.pagestree.style.landingpage.edited");
+		homePageEditedExpand = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.new_home_Content_Page_name_edited_expand");
+		menStylesForWinterEditedPageName = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.pagestree.articlesfolder.menstylesforwinter.page.edited");
+		
 
+		rightclickEditOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.edit.option");
 		rightclickViewOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.view.option");
+		rightclickNewContentOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.Content.option");
+		rightclickNewFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.folder.option");
+		rightclickCutOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.cut.option");
 		rightclickCopyOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.copy.option");
+		rightclickRequestPublishOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.requestpublish.option");
 		rightclickDependenciesOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.dependenciesoption");
-		technologyLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("dashboard.pagestree.technology.landingpage");
+		rightclickHistoryOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.history.option");
+		styleLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("dashboard.pagestree.style.landingpage");
 		rightclickDuplicateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.duplicate.option");
-		rightclickRenameFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.rename.folder.option");
 		articlesFolder = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.articlesfolder");
 		articlesFolder2017 = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -103,6 +132,10 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				.getProperty("dashboard.articles.folder.2017.1");
 		articlesFolderMenStylesForWinter = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.articles.folder.2017.1.menstylesforwinter");
+		createFormFrameElementCss = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("complexscenarios.general.createformframe");
+		rightClickOptions = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.list.all.options");
 		newUserFirstNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("general.users.firstname");
 		newUserLastNameId = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -115,7 +148,7 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		newUserPasswordVerificationId = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("general.users.passwordVerification");
 		newUserUserNameCreatedXpath = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("general.users.reviewerusernamecreated");
+				.getProperty("general.users.authorusernamecreated");
 		crafterLogo = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("users.crafterlogo");
 		expandPagesTree = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.expand_Pages_Tree");
@@ -126,8 +159,8 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				.getProperty("home.createsitebutton");
 		addTouserIframe = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("complexscenarios.general.adduser.iframe");
-		editReviewerGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("groups.edit_reviewer_group_option");
+		editAuthorGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.edit_author_group_option");
 		groupsAddNewMembersCheckbox = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("groups.add_new_members_checkbox");
 		groupsAddNewMembersInput = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -141,20 +174,18 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		userOptions = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("dashboard.user_options");
 		userOptionsLogout = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.user_options_logout");
-		rightClickOptions = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("rightclick.list.all.options");
 
 	}
 
-	public void rightClickHome() {
+	public void rightClickHomePage() {
 		this.driverManager.waitUntilPageLoad();
 		this.driverManager.waitUntilSidebarOpens();
 
 		this.driverManager.waitUntilFolderOpens("xpath", pagesTree);
 		this.driverManager.waitForAnimation();
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homeContent);
-		dashboardPage.getDriverManager().contextClick("xpath", homeContent, false);
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homePageEditedName);
+		dashboardPage.getDriverManager().contextClick("xpath", homePageEditedName, false);
 
 	}
 
@@ -163,18 +194,8 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		this.driverManager.waitUntilSidebarOpens();
 		this.driverManager.waitForAnimation();
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", technologyLandingpage);
-		dashboardPage.getDriverManager().contextClick("xpath", technologyLandingpage, false);
-		this.driverManager.waitForAnimation();
-	}
-
-	public void rightClickArticlesFolder1() {
-		this.driverManager.waitUntilPageLoad();
-		this.driverManager.waitUntilSidebarOpens();
-		this.driverManager.waitForAnimation();
-
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1);
-		dashboardPage.getDriverManager().contextClick("xpath", articlesFolder1, false);
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", menStylesEditedPageName);
+		dashboardPage.getDriverManager().contextClick("xpath", menStylesEditedPageName, false);
 		this.driverManager.waitForAnimation();
 	}
 
@@ -184,9 +205,16 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		this.driverManager.waitForAnimation();
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
-				articlesFolderMenStylesForWinter);
-		dashboardPage.getDriverManager().contextClick("xpath", articlesFolderMenStylesForWinter, false);
+				menStylesForWinterEditedPageName);
+		dashboardPage.getDriverManager().contextClick("xpath", menStylesForWinterEditedPageName, false);
 		this.driverManager.waitForAnimation();
+	}
+
+	public void verifyEditOptionIsPresent(String section) {
+		WebElement rightclickEditOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickEditOption);
+		Assert.assertTrue(rightclickEditOptionElement.isDisplayed(),
+				"ERROR: Right click Edit Option is not present on right click of " + section);
 	}
 
 	public void verifyViewOptionIsPresent(String section) {
@@ -196,11 +224,40 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				"ERROR: Right click View Option is not present on right click of " + section);
 	}
 
+	public void verifyNewContentOptionIsPresent(String section) {
+		WebElement rightclickNewContentOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewContentOption);
+		Assert.assertTrue(rightclickNewContentOptionElement.isDisplayed(),
+				"ERROR: Right click New Content Option is not present on right click of " + section);
+	}
+
+	public void verifyNewFolderOptionIsPresent(String section) {
+		WebElement rightclickNewFolderOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewFolderOption);
+		Assert.assertTrue(rightclickNewFolderOptionElement.isDisplayed(),
+				"ERROR: Right click New Folder Option is not present on right click of " + section);
+	}
+
+	public void verifyCutOptionIsPresent(String section) {
+		WebElement rightclickCutOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickCutOption);
+		Assert.assertTrue(rightclickCutOptionElement.isDisplayed(),
+				"ERROR: Right click Cut Option is not present on right click of " + section);
+	}
+
 	public void verifyCopyOptionIsPresent(String section) {
 		WebElement rightclickCopyOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 				rightclickCopyOption);
 		Assert.assertTrue(rightclickCopyOptionElement.isDisplayed(),
 				"ERROR: Right click Copy Option is not present on right click of " + section);
+	}
+	
+	public void verifyRequestPublishOptionIsPresent(String section) {
+		WebElement rightclickRequestPublishOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickRequestPublishOption);
+		Assert.assertTrue(rightclickRequestPublishOptionElement.isDisplayed(),
+				"ERROR: Right click Request Publish Option is not present on right click of " + section);
+
 	}
 
 	public void verifyDependenciesOptionIsPresent(String section) {
@@ -210,6 +267,13 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				"ERROR: Right click Dependencies Option is not present on right click of " + section);
 	}
 
+	public void verifyHistoryOptionIsPresent(String section) {
+		WebElement rightclickHistoryOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickHistoryOption);
+		Assert.assertTrue(rightclickHistoryOptionElement.isDisplayed(),
+				"ERROR: Right click History Option is not present on right click of " + section);
+	}
+
 	public void verifyDuplicateOptionIsPresent(String section) {
 		WebElement rightclickDuplicateOptionElement = this.driverManager
 				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDuplicateOption);
@@ -217,24 +281,22 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				"ERROR: Right click Duplicate Option is not present on right click of " + section);
 	}
 
-	public void verifyRenameFolderOptionIsPresent(String section) {
-		WebElement rightclickRenameFolderOptionElement = this.driverManager
-				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickRenameFolderOption);
-		Assert.assertTrue(rightclickRenameFolderOptionElement.isDisplayed(),
-				"ERROR: Right click Rename Folder Option is not present on right click of " + section);
-	}
-
 	public void step4() {
 		// Step 4 Right Right click on "Home" and verify options
-		this.rightClickHome();
+		this.rightClickHomePage();
 
 		driverManager.usingContextMenu(() -> {
-			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInHomePage = new LinkedList<String>();
-			rightClickOptionsListInHomePage.add(0, "View");
-			rightClickOptionsListInHomePage.add(1, "Copy");
-			rightClickOptionsListInHomePage.add(2, "Dependencies");
+			rightClickOptionsListInHomePage.add(0, "Edit");
+			rightClickOptionsListInHomePage.add(1, "View");
+			rightClickOptionsListInHomePage.add(2, "New Content");
+			rightClickOptionsListInHomePage.add(3, "New Folder");
+			rightClickOptionsListInHomePage.add(4, "Cut");
+			rightClickOptionsListInHomePage.add(5, "Copy");
+			rightClickOptionsListInHomePage.add(6, "Request Publish");
+			rightClickOptionsListInHomePage.add(7, "Dependencies");
+			rightClickOptionsListInHomePage.add(8, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -249,27 +311,40 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 			}
 
 			String section = "Step 4 Right Click on 'Home'";
-			
+
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
+			verifyRequestPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
+			this.driverManager.waitUntilSidebarOpens();
 		});
 	}
 
 	public void step6() {
 		// Step 6 Right click on any Category Landing page and verify options
 		this.rightClickCategoryLandingPage();
-
+		
 		driverManager.usingContextMenu(() -> {
-			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInCategoryLandingPage = new LinkedList<String>();
-			rightClickOptionsListInCategoryLandingPage.add(0, "View");
-			rightClickOptionsListInCategoryLandingPage.add(1, "Copy");
-			rightClickOptionsListInCategoryLandingPage.add(2, "Dependencies");
+			rightClickOptionsListInCategoryLandingPage.add(0, "Edit");
+			rightClickOptionsListInCategoryLandingPage.add(1, "View");
+			rightClickOptionsListInCategoryLandingPage.add(2, "New Content");
+			rightClickOptionsListInCategoryLandingPage.add(3, "New Folder");
+			rightClickOptionsListInCategoryLandingPage.add(4, "Cut");
+			rightClickOptionsListInCategoryLandingPage.add(5, "Copy");
+			rightClickOptionsListInCategoryLandingPage.add(6, "Duplicate");
+			rightClickOptionsListInCategoryLandingPage.add(7, "Request Publish");
+			rightClickOptionsListInCategoryLandingPage.add(8, "Dependencies");
+			rightClickOptionsListInCategoryLandingPage.add(9, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -285,9 +360,17 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 			}
 
 			String section = "Step 6 Right click on a 'Category Landing' page";
+
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
+			verifyRequestPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
@@ -304,9 +387,8 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder2017);
 	}
 
-
 	public void step10() {
-		// Step 11 Right click on any of the article (Men Styles For Winter)
+		// Step 10 Right click on any of the article (Men Styles For Winter)
 		this.driverManager.waitForAnimation();
 
 		this.rightClickArticlesFolderMenStylesForWinter();
@@ -315,9 +397,16 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInMenStylesForWinterPage = new LinkedList<String>();
-			rightClickOptionsListInMenStylesForWinterPage.add(0, "View");
-			rightClickOptionsListInMenStylesForWinterPage.add(1, "Copy");
-			rightClickOptionsListInMenStylesForWinterPage.add(2, "Dependencies");
+			rightClickOptionsListInMenStylesForWinterPage.add(0, "Edit");
+			rightClickOptionsListInMenStylesForWinterPage.add(1, "View");
+			rightClickOptionsListInMenStylesForWinterPage.add(2, "New Content");
+			rightClickOptionsListInMenStylesForWinterPage.add(3, "New Folder");
+			rightClickOptionsListInMenStylesForWinterPage.add(4, "Cut");
+			rightClickOptionsListInMenStylesForWinterPage.add(5, "Copy");
+			rightClickOptionsListInMenStylesForWinterPage.add(6, "Duplicate");
+			rightClickOptionsListInMenStylesForWinterPage.add(7, "Request Publish");
+			rightClickOptionsListInMenStylesForWinterPage.add(8, "Dependencies");
+			rightClickOptionsListInMenStylesForWinterPage.add(9, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -328,20 +417,101 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 				Assert.assertTrue(
 						element.getText().equals(rightClickOptionsListInMenStylesForWinterPage.get(currentIndex)),
 						"ERROR: Link Option: " + element.getText()
-								+ " is not in the correct order in the Men Styles For Witner page, check that the correct options are listed");
+								+ " is not in the correct order in the Men Styles For Winter page, check that the correct options are listed");
 				currentIndex++;
 			}
 
 			String section = "Step 11 Right click on folder articles -2017- 1- Men Styles For Winter;";
 
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
+			verifyRequestPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
 			this.driverManager.waitForAnimation();
 		});
-		
+
+	}
+
+	public void editHomePage() {
+		// Step 3 Edit the Home Page and Save
+
+		dashboardPage.rightClickEditOnAPresentPage(homeContent);
+
+		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
+
+			// creating random values for URL field and InternalName field
+
+			String randomInternalName = "HomePageEdited";
+
+			// Set basics fields of the new content created
+
+			dashboardPage.editInternalName(randomInternalName);
+		});
+
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", homePageEditedName);
+
+	}
+
+	public void editLandingPageStylePage() {
+
+		dashboardPage.rightClickEditOnAPresentPage(styleLandingpage);
+
+		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
+
+			// creating random values for URL field and InternalName field
+
+			String randomInternalName = "StylePageEdited";
+
+			// Set basics fields of the new content created
+
+			dashboardPage.editInternalName(randomInternalName);
+		});
+
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", menStylesEditedPageName);
+
+	}
+
+	public void editMenStylesForWinterPage() {
+
+		dashboardPage.rightClickEditOnAPresentPage(articlesFolderMenStylesForWinter);
+
+		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
+
+			// creating random values for URL field and InternalName field
+
+			String randomInternalName = "MenStylesForWinterEdited";
+
+			// Set basics fields of the new content created
+
+			dashboardPage.editInternalName(randomInternalName);
+		});
+
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("Xpath", menStylesForWinterEditedPageName);
+
+	}
+
+	public void expandHomeTree() {
+		this.driverManager.waitUntilSidebarOpens();
+		this.driverManager.waitForAnimation();
+
+		// Verify if the home tree is already expanded
+		if (!(this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", homePageEditedName)
+				.getAttribute("class").contains("open"))) {
+			this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", homePageEditedExpand)
+					.click();
+		}
+
 	}
 	
 	public void login(String user, String loginpassword) {
@@ -364,20 +534,20 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		usersPage.clickOnNewUser();
 
 		// Follow the form
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("author");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserLastNameId)
 				.sendKeys("Last Name");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserEmailId)
-				.sendKeys("reviewer@email.com");
+				.sendKeys("email@email.com");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("author");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("author");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordVerificationId)
-				.sendKeys("reviewer");
+				.sendKeys("author");
 
 		// Save Button
 		usersPage.clickOnSaveNewUser();
@@ -447,11 +617,11 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
 
-				editReviewerGroupOption);
+				editAuthorGroupOption);
 
 		this.driverManager
 
-				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editReviewerGroupOption)
+				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editAuthorGroupOption)
 
 				.click();
 
@@ -475,7 +645,7 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 
 				.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersInput)
 
-				.sendKeys("reviewer");
+				.sendKeys("author");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 
@@ -522,8 +692,8 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 	}
 
 	@Test(priority = 0)
-	public void verifyProperOptionDisplayedOnPagesStructureWithPublisherUser() {
-		
+	public void verifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingAuthorUser() {
+
 		this.login(userName, password);
 
 		logger.info("Adding New User");
@@ -540,7 +710,7 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 
 		this.driverManager.waitUntilSidebarOpens();
 
-		logger.info("Add previous created user to Reviewer Group");
+		logger.info("Add previous created user to Author Group");
 
 		this.addUserToAuthorGroup();
 
@@ -548,9 +718,9 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		logger.info("logout from Crafter");
 		this.logoutFromCrafter();
 
-		// login to application with reviewer user
-		logger.info("login to application with reviewer user");
-		loginPage.loginToCrafter("reviewer", "reviewer");
+		// login to application with author user
+		logger.info("login to application with author user");
+		loginPage.loginToCrafter("author", "author");
 
 		logger.info("Go to Preview Page");
 		this.homePage.goToPreviewPage();
@@ -569,16 +739,20 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		pagesTreeLinkElement.click();
 		this.driverManager.waitUntilFolderOpens("xpath", pagesTreeLink);
 
-		// Step 4 Right Right click on "Home" and verify options
+		this.editHomePage();
+
+		logger.info("Step 4 Right Right click on Home and verify options");
 		this.step4();
 
-		// Step 5 Click on the + of Home tree
-		this.dashboardPage.expandHomeTree();
+		logger.info("Step 5 Click on the + of Home tree and Edit Home Page");
+		this.expandHomeTree();
 
-		// Step 6 Right click on any Category Landing page and verify options
+		logger.info("Step 6 Right click on any Category Landing page style and verify options");
+		this.editLandingPageStylePage();
+
 		this.step6();
 
-		// Step 7 Click on the + of Articles tree
+		logger.info("Step 7 Click on the + of Articles tree");
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder);
 
 		this.driverManager.waitUntilContentTooltipIsHidden();
@@ -586,14 +760,17 @@ public class VerifyTheApplicationDisplaysTheProperAvailableOptionsWhenRightClick
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder);
 
-		// Step 8 Click on the + of folder 2017
+		logger.info("Step 8 Click on the + of folder 2017");
 		this.step8();
 
-		// Step 9 Click on the + of folder "1"
+		logger.info(
+				"Step 9 Click on the + of folder 1 and Edit then article page Men Styles For Winter under articles/2017/1");
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder1);
 
-		// Step 10 Right click on any of the article (Men Styles For Winter)
+		this.editMenStylesForWinterPage();
+
+		logger.info("Step 10 Right click on any of the article Men Styles For Winter");
 		this.step10();
 	}
 }
