+++ src/test/java/org/craftercms/studio/test/cases/sitedropdowntestcases/VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingDeveloperUser.java
@@ -20,8 +20,8 @@ import org.openqa.selenium.WebElement;
  * @author Juan Camacho A
  *
  */
-// Test Case ID:107
-public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThePagesStructureIsEditedUsingReviewerUser
+//Test Case Studio- Site Dropdown ID:17
+public class VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingDeveloperUser
 		extends BaseTest {
 
 	private String userName;
@@ -50,7 +50,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private String adminConsole;
 	private String siteconfigGroupsOption;
 	private String addTouserIframe;
-	private String editReviewerGroupOption;
+	private String editDeveloperGroupOption;
 	private String groupsAddNewMembersCheckbox;
 	private String groupsAddNewMembersInput;
 	private String groupsAddNewMembersAutocompleteOption1;
@@ -59,11 +59,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private String userOptions;
 	private String userOptionsLogout;
 
+	private String rightclickEditOption;
 	private String rightclickViewOption;
+	private String rightclickNewContentOption;
+	private String rightclickNewFolderOption;
+	private String rightclickDeleteOption;
+	private String rightclickChangeTemplateOption;
+	private String rightclickCutOption;
 	private String rightclickCopyOption;
+	private String rightclickDuplicateOption;
 	private String rightclickApproveAndPublishOption;
 	private String rightclickDependenciesOption;
+	private String rightclickHistoryOption;
 	private String styleLandingpage;
+	
 	private String articlesFolder;
 	private String articlesFolder2017;
 	private String articlesFolder1;
@@ -72,7 +81,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	private LinkedList<String> rightClickOptionsListInCategoryLandingPage;
 	private LinkedList<String> rightClickOptionsListInMenStylesForWinterPage;
 	private static Logger logger = LogManager.getLogger(
-			VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThePagesStructureIsEditedUsingReviewerUser.class);
+			VerifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingDeveloperUser.class);
 
 	@BeforeMethod
 	public void beforeTest() {
@@ -95,16 +104,33 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		menStylesForWinterEditedPageName = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.pagestree.articlesfolder.menstylesforwinter.page.edited");
 		
+
+		rightclickEditOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.edit.option");
 		rightclickViewOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.view.option");
+		rightclickNewContentOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.Content.option");
+		rightclickNewFolderOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.new.folder.option");
+		rightclickDeleteOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.delete.option");
+		rightclickChangeTemplateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.change.template.option");
+		rightclickCutOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.cut.option");
 		rightclickCopyOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.copy.option");
 		rightclickApproveAndPublishOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.approveandpublish.option");
 		rightclickDependenciesOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("rightclick.dependenciesoption");
+		rightclickHistoryOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.history.option");
 		styleLandingpage = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.pagestree.style.landingpage");
+		rightclickDuplicateOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("rightclick.duplicate.option");
 		articlesFolder = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.articlesfolder");
 		articlesFolder2017 = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -129,7 +155,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		newUserPasswordVerificationId = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("general.users.passwordVerification");
 		newUserUserNameCreatedXpath = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("general.users.reviewerusernamecreated");
+				.getProperty("general.users.developerusernamecreated");
 		crafterLogo = uiElementsPropertiesManager.getSharedUIElementsLocators().getProperty("users.crafterlogo");
 		expandPagesTree = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("dashboard.expand_Pages_Tree");
@@ -140,8 +166,8 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				.getProperty("home.createsitebutton");
 		addTouserIframe = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("complexscenarios.general.adduser.iframe");
-		editReviewerGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
-				.getProperty("groups.edit_reviewer_group_option");
+		editDeveloperGroupOption = uiElementsPropertiesManager.getSharedUIElementsLocators()
+				.getProperty("groups.edit_developer_group_option");
 		groupsAddNewMembersCheckbox = uiElementsPropertiesManager.getSharedUIElementsLocators()
 				.getProperty("groups.add_new_members_checkbox");
 		groupsAddNewMembersInput = uiElementsPropertiesManager.getSharedUIElementsLocators()
@@ -191,11 +217,53 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.waitForAnimation();
 	}
 
+	public void verifyEditOptionIsPresent(String section) {
+		WebElement rightclickEditOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickEditOption);
+		Assert.assertTrue(rightclickEditOptionElement.isDisplayed(),
+				"ERROR: Right click Edit Option is not present on right click of " + section);
+	}
+
 	public void verifyViewOptionIsPresent(String section) {
 		WebElement rightclickViewOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 				rightclickViewOption);
 		Assert.assertTrue(rightclickViewOptionElement.isDisplayed(),
 				"ERROR: Right click View Option is not present on right click of " + section);
+	}
+
+	public void verifyNewContentOptionIsPresent(String section) {
+		WebElement rightclickNewContentOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewContentOption);
+		Assert.assertTrue(rightclickNewContentOptionElement.isDisplayed(),
+				"ERROR: Right click New Content Option is not present on right click of " + section);
+	}
+
+	public void verifyNewFolderOptionIsPresent(String section) {
+		WebElement rightclickNewFolderOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickNewFolderOption);
+		Assert.assertTrue(rightclickNewFolderOptionElement.isDisplayed(),
+				"ERROR: Right click New Folder Option is not present on right click of " + section);
+	}
+
+	public void verifyDeleteOptionIsPresent(String section) {
+		WebElement rightclickDeleteOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDeleteOption);
+		Assert.assertTrue(rightclickDeleteOptionElement.isDisplayed(),
+				"ERROR: Right click Delete Option is not present on right click of " + section);
+	}
+
+	public void verifyChangeTemplateOptionIsPresent(String section) {
+		WebElement rightclickChangeTemplateOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickChangeTemplateOption);
+		Assert.assertTrue(rightclickChangeTemplateOptionElement.isDisplayed(),
+				"ERROR: Right click Change Template Option is not present on right click of " + section);
+	}
+
+	public void verifyCutOptionIsPresent(String section) {
+		WebElement rightclickCutOptionElement = this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
+				rightclickCutOption);
+		Assert.assertTrue(rightclickCutOptionElement.isDisplayed(),
+				"ERROR: Right click Cut Option is not present on right click of " + section);
 	}
 
 	public void verifyCopyOptionIsPresent(String section) {
@@ -219,6 +287,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				"ERROR: Right click Dependencies Option is not present on right click of " + section);
 	}
 
+	public void verifyHistoryOptionIsPresent(String section) {
+		WebElement rightclickHistoryOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickHistoryOption);
+		Assert.assertTrue(rightclickHistoryOptionElement.isDisplayed(),
+				"ERROR: Right click History Option is not present on right click of " + section);
+	}
+
+	public void verifyDuplicateOptionIsPresent(String section) {
+		WebElement rightclickDuplicateOptionElement = this.driverManager
+				.driverWaitUntilElementIsPresentAndDisplayed("xpath", rightclickDuplicateOption);
+		Assert.assertTrue(rightclickDuplicateOptionElement.isDisplayed(),
+				"ERROR: Right click Duplicate Option is not present on right click of " + section);
+	}
+
 	public void step4() {
 		// Step 4 Right Right click on "Home" and verify options
 		this.rightClickHomePage();
@@ -226,11 +308,16 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		driverManager.usingContextMenu(() -> {
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInHomePage = new LinkedList<String>();
-			
-			rightClickOptionsListInHomePage.add(0, "View");
-			rightClickOptionsListInHomePage.add(1, "Copy");
-			rightClickOptionsListInHomePage.add(2, "Approve & Publish");
-			rightClickOptionsListInHomePage.add(3, "Dependencies");
+			rightClickOptionsListInHomePage.add(0, "Edit");
+			rightClickOptionsListInHomePage.add(1, "View");
+			rightClickOptionsListInHomePage.add(2, "New Content");
+			rightClickOptionsListInHomePage.add(3, "New Folder");
+			rightClickOptionsListInHomePage.add(4, "Change Template");
+			rightClickOptionsListInHomePage.add(5, "Cut");
+			rightClickOptionsListInHomePage.add(6, "Copy");
+			rightClickOptionsListInHomePage.add(7, "Approve & Publish");
+			rightClickOptionsListInHomePage.add(8, "Dependencies");
+			rightClickOptionsListInHomePage.add(9, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -246,16 +333,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 			String section = "Step 4 Right Click on 'Home'";
 
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
 			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
-			
+			verifyHistoryOptionIsPresent(section);
+
 			this.driverManager.getDriver().navigate().refresh();
-			this.driverManager.waitUntilSidebarOpens();
 			this.driverManager.waitForAnimation();
-			
-	
+			this.driverManager.waitUntilSidebarOpens();
 		});
 	}
 
@@ -266,12 +357,18 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		driverManager.usingContextMenu(() -> {
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInCategoryLandingPage = new LinkedList<String>();
-			
-			rightClickOptionsListInCategoryLandingPage.add(0, "View");
-			rightClickOptionsListInCategoryLandingPage.add(1, "Copy");
-			rightClickOptionsListInCategoryLandingPage.add(2, "Approve & Publish");
-			rightClickOptionsListInCategoryLandingPage.add(3, "Dependencies");
-		
+			rightClickOptionsListInCategoryLandingPage.add(0, "Edit");
+			rightClickOptionsListInCategoryLandingPage.add(1, "View");
+			rightClickOptionsListInCategoryLandingPage.add(2, "New Content");
+			rightClickOptionsListInCategoryLandingPage.add(3, "New Folder");
+			rightClickOptionsListInCategoryLandingPage.add(4, "Delete");
+			rightClickOptionsListInCategoryLandingPage.add(5, "Change Template");
+			rightClickOptionsListInCategoryLandingPage.add(6, "Cut");
+			rightClickOptionsListInCategoryLandingPage.add(7, "Copy");
+			rightClickOptionsListInCategoryLandingPage.add(8, "Duplicate");
+			rightClickOptionsListInCategoryLandingPage.add(9, "Approve & Publish");
+			rightClickOptionsListInCategoryLandingPage.add(10, "Dependencies");
+			rightClickOptionsListInCategoryLandingPage.add(11, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -288,13 +385,20 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 			String section = "Step 6 Right click on a 'Category Landing' page";
 
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyDeleteOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
 			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
+			verifyHistoryOptionIsPresent(section);
 
 			this.driverManager.getDriver().navigate().refresh();
-			this.driverManager.waitUntilSidebarOpens();
 			this.driverManager.waitForAnimation();
 		});
 	}
@@ -319,11 +423,18 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 			
 			logger.info("Checking that only the expected options are listed");
 			rightClickOptionsListInMenStylesForWinterPage = new LinkedList<String>();
-
-			rightClickOptionsListInMenStylesForWinterPage.add(0, "View");
-			rightClickOptionsListInMenStylesForWinterPage.add(1, "Copy");
-			rightClickOptionsListInMenStylesForWinterPage.add(2, "Approve & Publish");
-			rightClickOptionsListInMenStylesForWinterPage.add(3, "Dependencies");
+			rightClickOptionsListInMenStylesForWinterPage.add(0, "Edit");
+			rightClickOptionsListInMenStylesForWinterPage.add(1, "View");
+			rightClickOptionsListInMenStylesForWinterPage.add(2, "New Content");
+			rightClickOptionsListInMenStylesForWinterPage.add(3, "New Folder");
+			rightClickOptionsListInMenStylesForWinterPage.add(4, "Delete");
+			rightClickOptionsListInMenStylesForWinterPage.add(5, "Change Template");
+			rightClickOptionsListInMenStylesForWinterPage.add(6, "Cut");
+			rightClickOptionsListInMenStylesForWinterPage.add(7, "Copy");
+			rightClickOptionsListInMenStylesForWinterPage.add(8, "Duplicate");
+			rightClickOptionsListInMenStylesForWinterPage.add(9, "Approve & Publish");
+			rightClickOptionsListInMenStylesForWinterPage.add(10, "Dependencies");
+			rightClickOptionsListInMenStylesForWinterPage.add(11, "History");
 
 			List<WebElement> rightClickOptionsList = this.driverManager.getDriver()
 					.findElements(By.xpath(rightClickOptions));
@@ -340,19 +451,28 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 			String section = "Step 11 Right click on folder articles -2017- 1- Men Styles For Winter;";
 
+			verifyEditOptionIsPresent(section);
 			verifyViewOptionIsPresent(section);
+			verifyNewContentOptionIsPresent(section);
+			verifyNewFolderOptionIsPresent(section);
+			verifyDeleteOptionIsPresent(section);
+			verifyChangeTemplateOptionIsPresent(section);
+			verifyCutOptionIsPresent(section);
 			verifyCopyOptionIsPresent(section);
+			verifyDuplicateOptionIsPresent(section);
 			verifyApproveAndPublishOptionIsPresent(section);
 			verifyDependenciesOptionIsPresent(section);
-			
+			verifyHistoryOptionIsPresent(section);
+
 			this.driverManager.getDriver().navigate().refresh();
+			this.driverManager.waitForAnimation();
 		});
 
 	}
 
 	public void editHomePage() {
-		// Edit the Home Page and Save
-		logger.info("Edit the Home Page and Save");
+		// Step 3 Edit the Home Page and Save
+
 		dashboardPage.rightClickEditOnAPresentPage(homeContent);
 
 		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
@@ -372,8 +492,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	}
 
 	public void editLandingPageStylePage() {
-		
-		logger.info("Edit the Styles Page and Save");
+
 		dashboardPage.rightClickEditOnAPresentPage(styleLandingpage);
 
 		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
@@ -393,9 +512,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	}
 
 	public void editMenStylesForWinterPage() {
-		
-		logger.info("Edit the Men Styles For Winter Page and Save");
-		
+
 		dashboardPage.rightClickEditOnAPresentPage(articlesFolderMenStylesForWinter);
 
 		driverManager.usingCrafterForm("cssSelector", createFormFrameElementCss, () -> {
@@ -447,7 +564,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		usersPage.clickOnNewUser();
 
 		// Follow the form
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserFirstNameId).sendKeys("developer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserLastNameId)
 				.sendKeys("Last Name");
@@ -455,12 +572,12 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserEmailId)
 				.sendKeys("email@email.com");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserUserNameId).sendKeys("developer");
 
-		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("reviewer");
+		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordId).sendKeys("developer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", newUserPasswordVerificationId)
-				.sendKeys("reviewer");
+				.sendKeys("developer");
 
 		// Save Button
 		usersPage.clickOnSaveNewUser();
@@ -501,7 +618,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 					"Site creation process is taking too long time and the element was not found");
 	}
 	
-	public void addUserToReviewerGroup() {
+	public void addUserToDeveloperGroup() {
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", adminConsole);
 
@@ -530,11 +647,11 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath",
 
-				editReviewerGroupOption);
+				editDeveloperGroupOption);
 
 		this.driverManager
 
-				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editReviewerGroupOption)
+				.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", editDeveloperGroupOption)
 
 				.click();
 
@@ -558,7 +675,7 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 				.driverWaitUntilElementIsPresentAndDisplayed("xpath", groupsAddNewMembersInput)
 
-				.sendKeys("reviewer");
+				.sendKeys("developer");
 
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath",
 
@@ -605,8 +722,8 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 	}
 
 	@Test(priority = 0)
-	public void verifyProperOptionDisplayedOnPagesStructureWithReviewerUser() {
-	
+	public void verifyRightClickOptionsOfAnEditedPageUnderPageStructureUsingDeveloperUser() {
+
 		this.login(userName, password);
 
 		logger.info("Adding New User");
@@ -623,10 +740,23 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 
 		this.driverManager.waitUntilSidebarOpens();
 
-		logger.info("Add previous created user to Reviewer Group");
+		logger.info("Add previous created user to Developer Group");
 
-		this.addUserToReviewerGroup();
-		
+		this.addUserToDeveloperGroup();
+
+		// logout from Crafter
+		logger.info("logout from Crafter");
+		this.logoutFromCrafter();
+
+		// login to application with developer user
+		logger.info("login to application with developer user");
+		loginPage.loginToCrafter("developer", "developer");
+
+		logger.info("Go to Preview Page");
+		this.homePage.goToPreviewPage();
+
+		this.driverManager.waitForAnimation();
+
 		// Expand the site bar
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayed("xpath", siteDropdownElementXPath);
 		
@@ -638,15 +768,21 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 				pagesTreeLink);
 		pagesTreeLinkElement.click();
 		this.driverManager.waitUntilFolderOpens("xpath", pagesTreeLink);
-		
+
 		this.editHomePage();
-		
-		logger.info("Click on the + of Home tree and Edit Home Page");
+
+		logger.info("Step 4 Right Right click on Home and verify options");
+		this.step4();
+
+		logger.info("Step 5 Click on the + of Home tree and Edit Home Page");
 		this.expandHomeTree();
-		
+
+		logger.info("Step 6 Right click on any Category Landing page style and verify options");
 		this.editLandingPageStylePage();
-		
-		logger.info("Click on the + of Articles tree");
+
+		this.step6();
+
+		logger.info("Step 7 Click on the + of Articles tree");
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder);
 
 		this.driverManager.waitUntilContentTooltipIsHidden();
@@ -654,36 +790,17 @@ public class VerifyAppDisplaysTheProperOptionsUsingRightClickWhenAnyElementOfThe
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder);
 
-		logger.info("Click on the + of folder 2017");
+		logger.info("Step 8 Click on the + of folder 2017");
 		this.step8();
 
 		logger.info(
-				"Click on the + of folder 1 and Edit then article page Men Styles For Winter under articles/2017/1");
+				"Step 9 Click on the + of folder 1 and Edit then article page Men Styles For Winter under articles/2017/1");
 		this.driverManager.driverWaitUntilElementIsPresentAndDisplayedAndClickable("xpath", articlesFolder1).click();
 		this.driverManager.waitUntilFolderOpens("xpath", articlesFolder1);
 
 		this.editMenStylesForWinterPage();
-		
-		// logout from Crafter
-		logger.info("logout from Crafter");
-		this.logoutFromCrafter();
 
-		// login to application with reviewer user
-		logger.info("login to application with reviewer user");
-		loginPage.loginToCrafter("reviewer", "reviewer");
-
-		logger.info("Go to Preview Page");
-		this.homePage.goToPreviewPage();
-
-		this.driverManager.waitForAnimation();
-
-		logger.info("Verifications of Step 4 Right Right click on Home and verify options");
-		this.step4();
-
-		logger.info("Verifications of Step 6 Right click on any Category Landing page style and verify options");
-		this.step6();
-
-		logger.info("Verifications of Step 10 Right click on any of the article, using Men Styles For Winter and verify options");
+		logger.info("Step 10 Right click on any of the article Men Styles For Winter");
 		this.step10();
 	}
 }
